Migrate testing framework and improve ephemeral test
