+++ internal/provider/resource_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
 
 func CheckJSONEqual(attr, expect string) resource.CheckResourceAttrWithFunc {
